Moved components from subfolder to main component folder. Also changed to use  functional component
